cosmology

+++ src/axioms/mathematics.h
@@ -24,7 +24,7 @@ static inline int cyclic_embind(int lower_bound, int value, int upper_bound) {
     return value;
 }
 
-static inline int cyclic_embind_flt(float lower_bound, float value, float upper_bound) {
+static inline float cyclic_embind_flt(float lower_bound, float value, float upper_bound) {
     if (value > upper_bound) value = lower_bound;
     if (value < lower_bound) value = upper_bound;
     return value;
